requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
loky==2.5.1
lxml==4.3.4
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
XlsxWriter==1.1.8
